# Improve performance of http access logs