@liggitt 2020-06-23
/test pull-kubernetes-verify

master verify is failing on go1.14 fmt/vet/staticcheck issues, fixed in https://github.com/kubernetes/kubernetes/pull/92438